Remove mysql exception handling.
When a mysql exception happens, there isn't really anything we can do.
The script should really just crash.

This is not documented, but connection objects are closed automatically
when they are garbage collected[0], so the "con.close if con" lines
are unnecessary. The mysql2 gem, which we may switch to at some point,
also has this behavior, and it is actually documented[1].

[0] https://github.com/luislavena/mysql-gem/blob/master/ext/mysql_api/mysql.c#L86
[1] http://rubydoc.info/gems/mysql2/Mysql2/Client#close-instance_method