Add migration of annotation links (which become generic "links" in the new database).
Links will automatically be migrated during a normal migration, but in
order to complete the operation the "resolve_links" task must be run. It
is possible that this will be run automatically at the end of a
migration at some point, but at this point we leave it as a separate
step.